Continuing to add core logic

+++ Position.cpp
@@ -2,44 +2,44 @@
 
 Position::Position()
 {
-    x_coord = 0;
-    y_coord = 0;
-    z_coord = 0;
+    x_coord = 0.0;
+    y_coord = 0.0;
+    z_coord = 0.0;
 }
 
-Position::Position(int x, int y, int z)
+Position::Position(float x, float y, float z)
 {
     x_coord = x;
     y_coord = y;
     z_coord = z;
 }
 
-void Position::setXCoord(int x)
+void Position::setXCoord(float x)
 {
     x_coord = x;
 }
 
-void Position::setYCoord(int y)
+void Position::setYCoord(float y)
 {
     y_coord = y;
 }
 
-void Position::setZCoord(int z)
+void Position::setZCoord(float z)
 {
     z_coord = z;
 }
 
-int Position::getXCoord()
+float Position::getXCoord()
 {
     return x_coord;
 }
 
-int Position::getYCoord()
+float Position::getYCoord()
 {
     return y_coord;
 }
 
-int Position::getZCoord()
+float Position::getZCoord()
 {
     return z_coord;
 }
@@ -58,7 +58,7 @@ void Position::subPositionFromThis(Position p)
     z_coord -= p.z_coord;
 }
 
-int Position::posMagnitudeSquared()
+float Position::posMagnitudeSquared()
 {
-    return pow(x_coord, 2) + pow(y_coord, 2) + pow(z_coord, 2);
+    return pow(x_coord, 2.0) + pow(y_coord, 2.0) + pow(z_coord, 2.0);
 }
